--- app/controllers/public_referrals/referrers_controller.rb
@@ -1,6 +1,0 @@
-module PublicReferrals
-  class ReferrersController < Referrals::ReferrersController
-    def personal_details
-    end
-  end
-end
