+++ config/locales/en.yml
@@ -250,22 +250,25 @@ en:
           attributes:
             teacher_role_complete:
               inclusion: Select yes if you’ve completed this section
-
-        referrer_job_title_form:
+        "referrals/referrer/job_title_form":
           attributes:
             job_title:
               blank: Enter your job title
-        referrer_phone_form:
+        "referrals/referrer/phone_form":
           attributes:
             phone:
               blank: Enter your phone number
               invalid: Enter a phone number, like 01632 960 001, 07700 900 982 or +44 808 157 0192
-        "referrals/referrer_name_form":
+        "referrals/referrer/name_form":
           attributes:
             first_name:
               blank: Enter your first name
             last_name:
               blank: Enter your last name
+        "referrals/referrer/check_answers_form":
+          attributes:
+            complete:
+              inclusion: Select yes if you’ve completed this section
         organisation_form:
           attributes:
             complete:
@@ -300,10 +303,6 @@ en:
           attributes:
             previous_misconduct_reported:
               inclusion: Let us know if there has been any previous misconduct, disciplinary action or complaints
-        referrer_form:
-          attributes:
-            complete:
-              inclusion: Select yes if you’ve completed this section
         "referrals/allegation/details_form":
           attributes:
             allegation_format:
