+++ spec/system/referrals/user_adds_their_details_spec.rb
@@ -90,7 +90,7 @@ RSpec.feature "Employer Referral: About You", type: :system do
 
   def then_i_am_on_the_your_name_page
     expect(page).to have_current_path(
-      "/referrals/#{@referral.id}/referrer-name/edit",
+      "/referrals/#{@referral.id}/referrer/name/edit",
       ignore_query: true
     )
     expect(page).to have_title(
@@ -101,7 +101,7 @@ RSpec.feature "Employer Referral: About You", type: :system do
 
   def then_i_see_the_job_title_page
     expect(page).to have_current_path(
-      "/referrals/#{Referral.last.id}/referrer-job-title/edit"
+      "/referrals/#{Referral.last.id}/referrer/job-title/edit"
     )
     expect(page).to have_title(
       "Your job title - Your details - Refer serious misconduct by a teacher in England - GOV.UK"
@@ -132,7 +132,9 @@ RSpec.feature "Employer Referral: About You", type: :system do
   end
 
   def then_i_see_the_referrer_check_your_answers_page
-    expect(page).to have_current_path("/referrals/#{@referral.id}/referrer")
+    expect(page).to have_current_path(
+      "/referrals/#{@referral.id}/referrer/check-answers/edit"
+    )
     expect(page).to have_title(
       "Your details - Refer serious misconduct by a teacher in England - GOV.UK"
     )
@@ -145,7 +147,7 @@ RSpec.feature "Employer Referral: About You", type: :system do
 
   def then_i_see_the_what_is_your_phone_number_page
     expect(page).to have_current_path(
-      "/referrals/#{@referral.id}/referrer-phone/edit"
+      "/referrals/#{@referral.id}/referrer/phone/edit"
     )
     expect(page).to have_title(
       "Your phone number - Your details - Refer serious misconduct by a teacher in England - GOV.UK"
